[SID:3755] fix(build): publishHistorySenderLabel — page.tsx named export 위반 수정
카디르 QA 지적(2026-09-09) — organization/events/page.tsx에 둔 named export
publishHistorySenderLabel이 Next App Router의 page.tsx export 화이트리스트
(metadata/generateMetadata/revalidate 등)를 위반해 "is not a valid Page
export field"로 next build가 실패했다(tsc/vitest는 이 층을 못 잡음, PR CI의
smoke-test에서만 드러남). 헬퍼를 member-display.ts로 이관 — 테스트도 함께
이동(publish-history-sender-label.test.ts 삭제, member-display.test.ts에
병합). next build 재현+통과 확認.
